stage2: Handshake and encrypted streams
Implements a challenge-response authentication handshake using RSA
negotiating a secret and initialization vector for a secret channel
using AES.

The !authenticate command (followed by the handshake) supersedes !login,
but is split logically as we are not permitted to use our own DTOs for
the handshake.
Therefore, the handshake is done seperately, resulting in object streams
for input and output, both backed by cipher streams that perform AES
transformations.

Server architecture changes quite a bit, as the server's main thread now
handles dispatch of threads for handshake and session handling upon
accept().
Users are tied to Sessions (and vice versa) 1:n, allowing for one user
to connect from multiple endpoints (even with different private keys,
though this has to be handled out-of-band).

Possible Issues:
 * Missing cleanup procedures. E.g. what happens when a user logs out,
   logs out of the last session or the server shuts down? Not sure this
   is all covered.
 * Handling of login could have had side-effects I did not consider.
   Similarily, logout may be handled inappropriately.
 * Side-effects on UDP and/or nameserver functionality.